Renamed sever_store crate to swim_persistence.
Also moved it under runtime rather than server.
