+++ runtime/swim_persistence/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "server_store"
+name = "swim_persistence"
 version = "0.1.0"
 edition = "2021"
