+++ swim_client/test_server/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 futures = "0.3.4"
 swim_model = { path = "../../api/swim_model" }
 swim_server = { path = "../../swim_server" }
-server_store = { path = "../../swim_server/server_store" }
+swim_persistence = { path = "../../runtime/swim_persistence" }
 url = "2.1.1"
 tracing = "0.1.16"
 tracing-futures = "0.2.4"
