+++ swim_server/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 [features]
 default = []
 debug = ["log_verbose"]
-persistence = ["server_store/rocks", "rocksdb"]
+persistence = ["swim_persistence/rocks", "rocksdb"]
 log_verbose = []
 
 [target.'cfg(log_verbose)'.dependencies]
@@ -53,7 +53,7 @@ lazy_static = "1.4.0"
 percent-encoding = "2.1.0"
 thiserror = "1.0"
 crossbeam-queue = "0.3"
-server_store = { path = "server_store"}
+swim_persistence = { path = "../runtime/swim_persistence" }
 serde = { version = "1.0.124", features = ["derive", "rc"] }
 bincode = "1.3.2"
 rocksdb = { version = "0.19.0", optional = true }
