+++ swim_server/src/agent/context/mod.rs
@@ -18,7 +18,6 @@ use crate::meta::MetaContext;
 use crate::scheduler::SchedulerContext;
 use futures::future::BoxFuture;
 use futures::Stream;
-use server_store::agent::NodeStore;
 use std::collections::HashMap;
 use std::future::Future;
 use std::sync::Arc;
@@ -26,6 +25,7 @@ use swim_async_runtime::time::clock::Clock;
 use swim_client::interface::ClientContext;
 use swim_metrics::NodeMetricAggregator;
 use swim_model::path::Path;
+use swim_persistence::agent::NodeStore;
 use swim_runtime::routing::Router;
 use swim_utilities::routing::uri::RelativeUri;
 use swim_utilities::time::AtomicInstant;
