+++ swim_server/src/agent/context/tests.rs
@@ -19,9 +19,6 @@ use crate::interface::ServerDownlinksConfig;
 use crate::meta::meta_context_sink;
 use crate::routing::TopLevelServerRouterFactory;
 use futures::future::BoxFuture;
-use server_store::agent::mock::MockNodeStore;
-use server_store::agent::SwimNodeStore;
-use server_store::plane::mock::MockPlaneStore;
 use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::sync::Arc;
@@ -30,6 +27,9 @@ use swim_async_runtime::time::clock::Clock;
 use swim_client::downlink::Downlinks;
 use swim_client::interface::ClientContext;
 use swim_client::router::ClientConnectionFactory;
+use swim_persistence::agent::mock::MockNodeStore;
+use swim_persistence::agent::SwimNodeStore;
+use swim_persistence::plane::mock::MockPlaneStore;
 use swim_runtime::configuration::DownlinkConnectionsConfig;
 use swim_runtime::error::{ResolutionError, RouterError};
 use swim_runtime::routing::{Route, Router, RoutingAddr};
