+++ swim_server/src/agent/dispatch/tests/mock.rs
@@ -23,9 +23,6 @@ use crate::agent::{AttachError, Eff, LaneIo};
 use futures::future::BoxFuture;
 use futures::FutureExt;
 use parking_lot::Mutex;
-use server_store::agent::mock::MockNodeStore;
-use server_store::agent::SwimNodeStore;
-use server_store::plane::mock::MockPlaneStore;
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
 use std::convert::TryFrom;
@@ -35,6 +32,9 @@ use swim_metrics::config::MetricAggregatorConfig;
 use swim_metrics::{MetaPulseLanes, NodeMetricAggregator};
 use swim_model::path::RelativePath;
 use swim_model::Value;
+use swim_persistence::agent::mock::MockNodeStore;
+use swim_persistence::agent::SwimNodeStore;
+use swim_persistence::plane::mock::MockPlaneStore;
 use swim_runtime::compat::{Operation, RequestMessage};
 use swim_runtime::error::{ConnectionDropped, ResolutionError, RouterError};
 use swim_runtime::routing::{Route, Router, RoutingAddr, TaggedEnvelope, TaggedSender};
