+++ swim_server/src/agent/lane/channels/task/tests.rs
@@ -33,9 +33,6 @@ use futures::future::{join, join3, ready, BoxFuture};
 use futures::stream::{once, BoxStream, FusedStream};
 use futures::{Future, FutureExt, Stream, StreamExt};
 use pin_utils::pin_mut;
-use server_store::agent::mock::MockNodeStore;
-use server_store::agent::SwimNodeStore;
-use server_store::plane::mock::MockPlaneStore;
 use std::collections::{HashMap, HashSet};
 use std::convert::{identity, TryFrom};
 use std::fmt::Debug;
@@ -57,6 +54,9 @@ use swim_metrics::uplink::{MetricBackpressureConfig, UplinkObserver, WarpUplinkP
 use swim_metrics::{AggregatorError, MetaPulseLanes, NodeMetricAggregator};
 use swim_model::path::RelativePath;
 use swim_model::Value;
+use swim_persistence::agent::mock::MockNodeStore;
+use swim_persistence::agent::SwimNodeStore;
+use swim_persistence::plane::mock::MockPlaneStore;
 use swim_runtime::compat::RequestMessage;
 use swim_runtime::error::{ConnectionDropped, ResolutionError, RouterError};
 use swim_runtime::routing::{Route, Router, RoutingAddr, TaggedEnvelope, TaggedSender};
