+++ swim_server/src/agent/lane/model/map/mod.rs
@@ -36,9 +36,6 @@ use futures::Stream;
 use futures::StreamExt;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
-use server_store::agent::lane::map::MapStoreEvent;
-use server_store::agent::lane::map::{MapDataModel, MapLaneStoreIo};
-use server_store::agent::NodeStore;
 use std::collections::HashMap;
 use std::fmt::Debug;
 use std::hash::Hash;
@@ -48,6 +45,9 @@ use stm::stm::{abort, left, right, Constant, Stm, VecStm, UNIT};
 use stm::transaction::{atomically, RetryManager, TransactionError, TransactionRunner};
 use stm::var::observer::{Observer, ObserverStream, ObserverSubscriber};
 use swim_form::structural::read::ReadError;
+use swim_persistence::agent::lane::map::MapStoreEvent;
+use swim_persistence::agent::lane::map::{MapDataModel, MapLaneStoreIo};
+use swim_persistence::agent::NodeStore;
 use swim_utilities::future::{FlatmapStream, SwimStreamExt, Transform};
 use swim_warp::map::MapUpdate;
 use tracing::{event, Level};
