+++ swim_server/src/agent/lane/model/map/tests.rs
@@ -19,12 +19,6 @@ use crate::agent::lane::model::DeferredSubscription;
 use crate::agent::lane::tests::ExactlyOnce;
 use futures::future::ready;
 use futures::{FutureExt, Stream, StreamExt};
-use server_store::agent::lane::error::StoreErrorHandler;
-use server_store::agent::lane::map::{MapDataModel, MapLaneStoreIo};
-use server_store::agent::lane::StoreIo;
-use server_store::agent::NodeStore;
-use server_store::plane::mock::MockPlaneStore;
-use server_store::{StoreEngine, StoreKey};
 use std::collections::HashMap;
 use std::num::NonZeroUsize;
 use std::ops::Deref;
@@ -33,6 +27,12 @@ use stm::stm::Stm;
 use stm::transaction::{atomically, TransactionRunner};
 use swim_form::Form;
 use swim_model::{Attr, Item, Value};
+use swim_persistence::agent::lane::error::StoreErrorHandler;
+use swim_persistence::agent::lane::map::{MapDataModel, MapLaneStoreIo};
+use swim_persistence::agent::lane::StoreIo;
+use swim_persistence::agent::NodeStore;
+use swim_persistence::plane::mock::MockPlaneStore;
+use swim_persistence::{StoreEngine, StoreKey};
 use swim_store::{serialize, EngineInfo, StoreError};
 use swim_utilities::algebra::non_zero_usize;
