+++ swim_server/src/agent/mod.rs
@@ -67,11 +67,6 @@ use futures::{FutureExt, Stream, StreamExt};
 use pin_utils::pin_mut;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
-use server_store::agent::lane::map::{MapDataModel, MapLaneStoreIo};
-use server_store::agent::lane::task::{NodeStoreErrors, NodeStoreTask};
-use server_store::agent::lane::value::{ValueDataModel, ValueLaneStoreIo};
-use server_store::agent::lane::StoreIo;
-use server_store::agent::NodeStore;
 use std::any::Any;
 use std::collections::HashMap;
 use std::error::Error;
@@ -85,6 +80,11 @@ use swim_client::interface::ClientContext;
 use swim_form::Form;
 use swim_model::path::{Path, RelativePath};
 use swim_model::Value;
+use swim_persistence::agent::lane::map::{MapDataModel, MapLaneStoreIo};
+use swim_persistence::agent::lane::task::{NodeStoreErrors, NodeStoreTask};
+use swim_persistence::agent::lane::value::{ValueDataModel, ValueLaneStoreIo};
+use swim_persistence::agent::lane::StoreIo;
+use swim_persistence::agent::NodeStore;
 use swim_runtime::compat::RequestMessage;
 use swim_runtime::routing::Router;
 use swim_utilities::future::SwimStreamExt;
