+++ swim_server/src/agent/tests/mod.rs
@@ -44,7 +44,6 @@ use crate::plane::provider::AgentProvider;
 use crate::routing::TopLevelServerRouterFactory;
 use futures::future::{join, BoxFuture};
 use futures::Stream;
-use server_store::agent::mock::MockNodeStore;
 use std::collections::HashMap;
 use std::fmt::Debug;
 use std::future::Future;
@@ -54,6 +53,7 @@ use swim_client::downlink::Downlinks;
 use swim_client::interface::ClientContext;
 use swim_client::router::ClientConnectionFactory;
 use swim_model::path::Path;
+use swim_persistence::agent::mock::MockNodeStore;
 use swim_runtime::configuration::DownlinkConnectionsConfig;
 use swim_utilities::algebra::non_zero_usize;
 use swim_utilities::routing::uri::RelativeUri;
