+++ swim_server/src/meta/info/tests.rs
@@ -24,7 +24,6 @@ use crate::plane::provider::AgentProvider;
 use crate::routing::TopLevelServerRouterFactory;
 use futures::future::BoxFuture;
 use futures::FutureExt;
-use server_store::agent::mock::MockNodeStore;
 use std::collections::{HashMap, HashSet};
 use std::convert::TryFrom;
 use std::sync::Arc;
@@ -35,6 +34,7 @@ use swim_client::router::ClientConnectionFactory;
 use swim_form::structural::read::ReadError;
 use swim_form::Form;
 use swim_model::record;
+use swim_persistence::agent::mock::MockNodeStore;
 use swim_runtime::configuration::DownlinkConnectionsConfig;
 use swim_runtime::error::{ConnectionDropped, ResolutionError, RouterError};
 use swim_runtime::routing::{Route, Router, RoutingAddr, TaggedEnvelope, TaggedSender};
