+++ swim_server/src/meta/log/tests.rs
@@ -27,7 +27,6 @@ use crate::plane::provider::AgentProvider;
 use crate::routing::TopLevelServerRouterFactory;
 use futures::future::BoxFuture;
 use futures::FutureExt;
-use server_store::agent::mock::MockNodeStore;
 use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::str::FromStr;
@@ -37,6 +36,7 @@ use swim_client::interface::ClientContext;
 use swim_client::router::ClientConnectionFactory;
 use swim_form::Form;
 use swim_model::Value;
+use swim_persistence::agent::mock::MockNodeStore;
 use swim_runtime::configuration::DownlinkConnectionsConfig;
 use swim_runtime::error::{ConnectionDropped, ResolutionError, RouterError};
 use swim_runtime::routing::{Route, Router, RoutingAddr, TaggedEnvelope, TaggedSender};
