+++ swim_server/src/plane/mod.rs
@@ -25,7 +25,6 @@ use futures::future::{join, BoxFuture};
 use futures::{select_biased, FutureExt, StreamExt};
 use futures_util::stream::TakeUntil;
 use pin_utils::pin_mut;
-use server_store::plane::PlaneStore;
 use std::any::Any;
 use std::collections::{HashMap, HashSet};
 use std::fmt::Debug;
@@ -34,6 +33,7 @@ use std::sync::{Arc, Weak};
 use swim_async_runtime::time::clock::Clock;
 use swim_client::interface::ClientContext;
 use swim_model::path::Path;
+use swim_persistence::plane::PlaneStore;
 use swim_runtime::error::{ConnectionDropped, NoAgentAtRoute, RouterError, Unresolvable};
 use swim_runtime::remote::RawOutRoute;
 use swim_runtime::routing::{CloseReceiver, RouterFactory, RoutingAddr, TaggedEnvelope};
