+++ swim_server/src/plane/spec/mod.rs
@@ -19,9 +19,9 @@ use crate::plane::lifecycle::PlaneLifecycle;
 use crate::plane::provider::AgentProvider;
 use crate::plane::{AgentRoute, BoxAgentRoute, PlaneSpec};
 use futures::Stream;
-use server_store::plane::PlaneStore;
 use std::fmt::Debug;
 use swim_async_runtime::time::clock::Clock;
+use swim_persistence::plane::PlaneStore;
 use swim_runtime::routing::{Router, TaggedEnvelope};
 use swim_utilities::routing::route_pattern::RoutePattern;
