+++ swim_server/src/plane/spec/tests.rs
@@ -23,10 +23,10 @@ use crate::plane::router::PlaneRouter;
 use crate::plane::spec::{PlaneBuilder, PlaneSpec, RouteSpec};
 use futures::future::{ready, BoxFuture, Ready};
 use futures::FutureExt;
-use server_store::agent::NodeStore;
-use server_store::plane::mock::MockPlaneStore;
 use std::time::Duration;
 use swim_async_runtime::time::clock::Clock;
+use swim_persistence::agent::NodeStore;
+use swim_persistence::plane::mock::MockPlaneStore;
 use swim_runtime::error::{ResolutionError, RouterError};
 use swim_runtime::routing::{Route, Router, RoutingAddr, TaggedEnvelope};
 use swim_utilities::routing::route_pattern::RoutePattern;
