+++ swim_server/src/plane/tests/fixture.rs
@@ -23,12 +23,12 @@ use futures::{FutureExt, StreamExt};
 use http::Uri;
 use parking_lot::Mutex;
 use pin_utils::pin_mut;
-use server_store::agent::SwimNodeStore;
-use server_store::plane::mock::MockPlaneStore;
 use std::any::Any;
 use std::sync::Arc;
 use std::time::Duration;
 use swim_async_runtime::time::clock::Clock;
+use swim_persistence::agent::SwimNodeStore;
+use swim_persistence::plane::mock::MockPlaneStore;
 use swim_runtime::routing::{Router, TaggedEnvelope};
 use swim_utilities::algebra::non_zero_usize;
 use swim_utilities::routing::uri::RelativeUri;
